chore:Move Area definition to domain package
