+++ agora/config.go
@@ -4,8 +4,8 @@ import (
 	"time"
 
 	"github.com/AgoraIO-Community/agora-rest-client-go/agora/auth"
+	"github.com/AgoraIO-Community/agora-rest-client-go/agora/domain"
 	"github.com/AgoraIO-Community/agora-rest-client-go/agora/log"
-	"github.com/AgoraIO-Community/agora-rest-client-go/agora/region"
 )
 
 type Config struct {
@@ -13,6 +13,6 @@ type Config struct {
 	HttpTimeout time.Duration
 	Credential  auth.Credential
 
-	RegionCode region.Area
+	DomainArea domain.Area
 	Logger     log.Logger
 }
