+++ agora/domain/region.go
@@ -1,0 +1,11 @@
+package domain
+
+type Area int
+
+const (
+	Unknown Area = iota
+	US
+	EU
+	AP
+	CN
+)
