--- agora/region/region.go
@@ -1,11 +1,0 @@
-package region
-
-type Area int
-
-const (
-	UnknownArea Area = iota
-	USArea
-	EUArea
-	APArea
-	CNArea
-)
