+++ agora/utils/utils.go
@@ -1,4 +1,4 @@
-package agora
+package utils
 
 import "reflect"
