+++ examples/cloudrecording/webrecording/service.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	"github.com/AgoraIO-Community/agora-rest-client-go/agora"
+	"github.com/AgoraIO-Community/agora-rest-client-go/agora/domain"
 	agoraLogger "github.com/AgoraIO-Community/agora-rest-client-go/agora/log"
-	"github.com/AgoraIO-Community/agora-rest-client-go/agora/region"
 	"github.com/AgoraIO-Community/agora-rest-client-go/examples/cloudrecording/base"
 	"github.com/AgoraIO-Community/agora-rest-client-go/services/cloudrecording"
 	cloudRecordingAPI "github.com/AgoraIO-Community/agora-rest-client-go/services/cloudrecording/api"
@@ -18,10 +18,10 @@ type Service struct {
 	base.Service
 }
 
-func NewService(region region.Area, appId, cname, uid string) *Service {
+func NewService(region domain.Area, appId, cname, uid string) *Service {
 	return &Service{
 		base.Service{
-			RegionArea: region,
+			DomainArea: region,
 			AppId:      appId,
 			Cname:      cname,
 			Uid:        uid,
@@ -35,7 +35,7 @@ func (s *Service) RunWebRecorder(storageConfig *cloudRecordingAPI.StorageConfig)
 	config := &agora.Config{
 		AppID:      s.AppId,
 		Credential: s.Credential,
-		RegionCode: s.RegionArea,
+		DomainArea: s.DomainArea,
 		Logger:     agoraLogger.NewDefaultLogger(agoraLogger.DebugLevel),
 	}
 
@@ -166,7 +166,7 @@ func (s *Service) RunWebRecorderAndRtmpPublish(storageConfig *cloudRecordingAPI.
 	config := &agora.Config{
 		AppID:      s.AppId,
 		Credential: s.Credential,
-		RegionCode: s.RegionArea,
+		DomainArea: s.DomainArea,
 		Logger:     agoraLogger.NewDefaultLogger(agoraLogger.DebugLevel),
 	}
